requirements.txt
# Requirements file (UTF-8 encoding)
flask==3.0.0
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0  # API client
colorama==0.4.6    # terminal colors
pytest==8.0.2
pytest-env==0.8.2 